feat!: move all exception class under Exception namespace
